Wedge6: Add methods that take Kokkos views instead of POD arrays.
