+++ include/master_element/MasterElement.h
@@ -399,143 +399,6 @@ public:
     double *result);
 };
 
-// Wedge 6 subcontrol volume
-class WedSCV : public MasterElement
-{
-public:
-  WedSCV();
-  virtual ~WedSCV();
-
-  const int * ipNodeMap(int ordinal = 0);
-
-  void determinant(
-    const int nelem,
-    const double *coords,
-    double *areav,
-    double * error );
-
-  void shape_fcn(
-    double *shpfc);
-
-  void shifted_shape_fcn(
-    double *shpfc);
-
-  void wedge_shape_fcn(
-    const int &npts,
-    const double *par_coord, 
-    double* shape_fcn);
-};
-
-// Wedge 6 subcontrol surface
-class WedSCS : public MasterElement
-{
-public:
-  WedSCS();
-  virtual ~WedSCS();
-
-  const int * ipNodeMap(int ordinal = 0);
-
-  void determinant(
-    const int nelem,
-    const double *coords,
-    double *areav,
-    double * error );
-
-  void grad_op(
-    const int nelem,
-    const double *coords,
-    double *gradop,
-    double *deriv,
-    double *det_j,
-    double * error );
-
-  void shifted_grad_op(
-    const int nelem,
-    const double *coords,
-    double *gradop,
-    double *deriv,
-    double *det_j,
-    double * error );
-
-  void wedge_derivative(
-    const int npts,
-    const double *intLoc,
-    double *deriv);
-
-  void face_grad_op(
-    const int nelem,
-    const int face_ordinal,
-    const double *coords,
-    double *gradop,
-    double *det_j,
-    double * error );
-
-  void shifted_face_grad_op(
-    const int nelem,
-    const int face_ordinal,
-    const double *coords,
-    double *gradop,
-    double *det_j,
-    double * error );
-
-  void gij(
-    const double *coords,
-    double *gupperij,
-    double *glowerij,
-    double *deriv);
-
-  const int * adjacentNodes();
-
-  int opposingNodes(
-    const int ordinal, const int node);
-
-  int opposingFace(
-    const int ordinal, const int node);
-  
-  void shape_fcn(
-    double *shpfc);
-
-  void shifted_shape_fcn(
-    double *shpfc);
-
-  double isInElement(
-    const double *elemNodalCoord,
-    const double *pointCoord,
-    double *isoParCoord);
-
-  void interpolatePoint(
-    const int &nComp,
-    const double *isoParCoord,
-    const double *field,
-    double *result);
-  
-  void wedge_shape_fcn(
-    const int &npts,
-    const double *par_coord, 
-    double* shape_fcn);
-
-  void general_face_grad_op(
-    const int face_ordinal,
-    const double *isoParCoord,
-    const double *coords,
-    double *gradop,
-    double *det_j,
-    double * error );
-
-  void sidePcoords_to_elemPcoords(
-    const int & side_ordinal,
-    const int & npoints,
-    const double *side_pcoords,
-    double *elem_pcoords);
-
-  // helper functions to isInElement
-  double parametric_distance( const double X, const double Y);
-  double parametric_distance( const std::vector<double> &x);
-
-  const int* side_node_ordinals(int sideOrdinal) final;
-
-};
-
 // 2D Quad 4 subcontrol volume
 class Quad2DSCV : public MasterElement
 {
